chore: move api to internal/web package
