+++ cmd/api/main.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/google/gops/agent"
 
-	"github.com/tsuru/rpaas-operator/api"
 	"github.com/tsuru/rpaas-operator/internal/config"
+	"github.com/tsuru/rpaas-operator/internal/web"
 )
 
 func main() {
@@ -23,7 +23,7 @@ func main() {
 	}
 	defer agent.Close()
 
-	a, err := api.New(nil)
+	a, err := web.New(nil)
 	if err != nil {
 		log.Fatalf("could not create RPaaS API: %v", err)
 	}
