+++ internal/web/extra_files.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package api
+package web
 
 import (
 	"fmt"
@@ -12,6 +12,7 @@ import (
 	"net/url"
 
 	"github.com/labstack/echo/v4"
+
 	"github.com/tsuru/rpaas-operator/internal/pkg/rpaas"
 )
