+++ internal/web/service_handlers.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package api
+package web
 
 import (
 	"bytes"
@@ -14,6 +14,7 @@ import (
 
 	"github.com/ajg/form"
 	"github.com/labstack/echo/v4"
+
 	"github.com/tsuru/rpaas-operator/internal/pkg/rpaas"
 )
